Major overhaul - portrait background
With the addition of several new background images (some from CK2+ and
others from unused African portrait pack assets), portrait background is
more dynamic than before.
For example, "military" background with the trebuchet is only shown when
in siege; portrait background changes in accordance with the geographic
feature of where the character is when leading armies.