Ocean portrait background restored
Ocean background became necessary because of the geography-dependent
code. The background is used for characters who are on ships or are in a
province with a port.
The frame was restored to where it is in vanilla background.dds; I had
to sacrifice the tavern background in exchange. I plan on sacrificing
the harem background if I get my mountain background.