Bolgar portrait overhaul
I was never satisfied with how the Bolgar portrait looked, and this one
looks much more appealing IMO. I had specifically wanted to use the
Cuman headgear since it's not used outside the Cuman portrait, but it
was hard to work with. The current version is much closer to what I had
originally in mind.

- Male beard: central asian -> cuman
- Male headgear and clothes: cuman -> steppe, steppe -> christian
turkish
- Female headgear and clothes: cuman -> steppe, steppe -> cuman